+++ src/Exception/UserException.php
@@ -4,14 +4,34 @@ declare(strict_types=1);
 
 namespace Yiisoft\ErrorHandler\Exception;
 
+use Attribute;
 use Exception;
+use ReflectionClass;
+use Throwable;
+
+use function count;
 
 /**
- * UserException is the base class for exceptions that are meant to be shown to end users.
- * Such exceptions are often caused by mistakes of end users.
+ * `UserException` is an exception and a class attribute that indicates
+ * the exception message is safe to display to end users.
+ *
+ * Usage:
+ * - throw directly (`throw new UserException(...)`) for explicit user-facing errors;
+ * - annotate any exception class with the `#[UserException]` attribute
+ *   to mark its messages as user-facing without extending this class.
  *
  * @final
  */
+#[Attribute(Attribute::TARGET_CLASS)]
 class UserException extends Exception
 {
+    public static function isUserException(Throwable $throwable): bool
+    {
+        if ($throwable instanceof self) {
+            return true;
+        }
+
+        $attributes = (new ReflectionClass($throwable))->getAttributes(self::class);
+        return count($attributes) > 0;
+    }
 }