+++ tests/Exception/UserException/NotFoundException.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\ErrorHandler\Tests\Exception\UserException;
+
+use Exception;
+use Yiisoft\ErrorHandler\Exception\UserException;
+
+#[UserException]
+final class NotFoundException extends Exception
+{
+}